requirements.txt

fastapi
uvicorn
pandas
sentence_transformers
pymongo
python-dotenv
